Proper CLI Project
Figured out how to use `npm prepare` to install things as a standalone project, in conjunction with using it as a global command.

Things are very hacked as of how it currently is, but that's always how it starts out. I really do need to complete the packaging of Region-Types, that's the main headache of everything, because you have to clone it as a submodule, as well as build it from source every time. Distributing it on npm is nice in that you get the "binary" version of the project, no matter how you wrote it. That's what makes it nice to have build steps I guess, it's easier to share the result of your code all together. It's not so easy to do this with the need of using Node APIs and LevelDB-Zlib though, there's no way to share it without installing those as well.

https://www.typescriptlang.org/docs/handbook/release-notes/typescript-5-2.html
https://docs.npmjs.com/cli/v10/using-npm/scripts
https://stackoverflow.com/questions/21363314/npm-install-dependencies-of-a-git-submodule